// src/types/ColorPalette.ts
export interface ColorPalette {
    backgroundPrimary: string;
    backgroundSecondary: string;
    backgroundTertiary: string;
    
    textPrimary: string;
    textSecondary: string;
    textTertiary: string;
    
    success: string;
    error: string;
    warning: string;
    info: string;
    
    borderPrimary: string;
    borderSecondary: string;
    
    iconPrimary: string;
    iconSecondary: string;
    
    shadowColor: string;
    
    accentPrimary: string;
    accentSecondary: string;
  }